fix(benchmark diffs): fix some minor issues
window_comination is partially fixed. It would require to reenable
selecting the root directory (useful for tests and unusual project
structures).

removed assertions in eq_node_cs because they trigger on false
positives, when enough nodes are stored and collisions start. Iwas not
getting the issue in tests because the inputs were not large enough...